added medicalClinic

+++ medical-clinic-FE/src/app/appointments/components/future-appointment/future-appointment.component.ts
@@ -5,6 +5,8 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
 import { SnackbarService } from '../../../guard/snackbar.service';
 import { Location } from '@angular/common';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-future-appointment',
@@ -18,7 +20,8 @@ export class FutureAppointmentComponent {
     private snackBar: MatSnackBar,
     private router: Router,
     private snackBarService: SnackbarService,
-    private location: Location
+    private location: Location,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -33,21 +36,30 @@ export class FutureAppointmentComponent {
   }
 
   deleteVisit(visit: VisitModel) {
-    const visitId = this.getVisitIdFromHours(visit.hours);
-    console.log('visitId ', visitId);
-    visit.hours[0].visitId;
-    this.appointmentsService.deleteVisit(visitId!).subscribe(
-      () => {
-        this.snackBarService.snackMessage('Wizyta usunięta poprawnie');
-        setTimeout(() => {
-          this.loadFutureVisits();
-        }, 2000);
-      },
-      error => {
-        console.error('Error:', error);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '250px',
+      data: { message: 'Chcesz usunąć tą wizytę?' },
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        const visitId = this.getVisitIdFromHours(visit.hours);
+        console.log('visitId ', visitId);
+
+        this.appointmentsService.deleteVisit(visitId!).subscribe(
+          () => {
+            this.snackBarService.snackMessage('Wizyta usunięta poprawnie');
+            window.location.reload();
+            setTimeout(() => {
+              this.loadFutureVisits();
+            }, 1500);
+          },
+          error => {
+            console.error('Error:', error);
+          }
+        );
       }
-    );
-    console.log('visitId2 ', visitId);
+    });
   }
   getVisitIdFromHours(hours: HoursModel[]): string | null {
     if (hours && hours.length > 0) {
